refactor(iocuak): move CreateInstanceTaskGraphExpandCommand
