+++ packages/iocuak/src/createInstanceTask/models/cuaktask/CreateInstanceTaskGraphExpandCommand.ts
@@ -1,9 +1,9 @@
 import { Task } from '@cuaklabs/cuaktask';
 
-import { TaskGraphExpandCommand } from '../../../common/models/cuaktask/TaskGraphExpandCommand';
 import { CreateInstanceTaskKind } from '../domain/CreateInstanceTaskKind';
 import { TaskKindType } from '../domain/TaskKindType';
 import { CreateInstanceTaskGraphExpandOperationContext } from './CreateInstanceTaskGraphExpandOperationContext';
+import { TaskGraphExpandCommand } from './TaskGraphExpandCommand';
 
 export type CreateInstanceTaskGraphExpandCommand = TaskGraphExpandCommand<
   CreateInstanceTaskGraphExpandOperationContext,
