+++ packages/iocuak/src/createInstanceTask/models/cuaktask/CreateInstanceTaskGraphExpandOperationContext.ts
@@ -1,9 +1,9 @@
 import * as cuaktask from '@cuaklabs/cuaktask';
 
-import { TaskGraphExpandOperationContext } from '../../../common/models/cuaktask/TaskGraphExpandOperationContext';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
 import { ReadOnlyLinkedList } from '../../../list/models/domain/ReadOnlyLinkedList';
 import { TaskKind } from '../domain/TaskKind';
+import { TaskGraphExpandOperationContext } from './TaskGraphExpandOperationContext';
 
 export interface CreateInstanceTaskGraphExpandOperationContext
   extends TaskGraphExpandOperationContext {
