+++ packages/iocuak/src/createInstanceTask/models/cuaktask/GetInstanceDependenciesTaskGraphExpandCommand.ts
@@ -1,9 +1,9 @@
 import { Task } from '@cuaklabs/cuaktask';
 
-import { TaskGraphExpandCommand } from '../../../common/models/cuaktask/TaskGraphExpandCommand';
 import { GetInstanceDependenciesTaskKind } from '../domain/GetInstanceDependenciesTaskKind';
 import { TaskKindType } from '../domain/TaskKindType';
 import { CreateInstanceTaskGraphExpandOperationContext } from './CreateInstanceTaskGraphExpandOperationContext';
+import { TaskGraphExpandCommand } from './TaskGraphExpandCommand';
 
 export type GetInstanceDependenciesTaskGraphExpandCommand =
   TaskGraphExpandCommand<
